Edits based on comments
