+++ qiskit/ignis/characterization/calibrations/__init__.py
@@ -17,6 +17,6 @@
 Calibration module
 """
 
-from .schedules import rabi_schedules, drag_schedules
+from .pulse_schedules import rabi_schedules, drag_schedules
 from .fitters import RabiFitter, DragFitter
 from .ibmq_utils import get_single_q_pulse, update_u_gates
